Added stub for mlt render
